+++ libraries/exoplayer_rtsp/src/main/java/androidx/media3/exoplayer/rtsp/reader/RtpH263Reader.java
@@ -15,6 +15,8 @@
  */
 package androidx.media3.exoplayer.rtsp.reader;
 
+import static androidx.media3.common.util.Assertions.checkNotNull;
+import static androidx.media3.common.util.Assertions.checkState;
 import static androidx.media3.common.util.Assertions.checkStateNotNull;
 
 import androidx.media3.common.C;
@@ -61,6 +63,12 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
   private boolean isKeyFrame;
   private boolean isOutputFormatSet;
   private long startTimeOffsetUs;
+  private long fragmentedSampleTimeUs;
+  /**
+   * Whether the first packet of a H263 frame is received, it mark the start of a H263 partition. A
+   * H263 frame can be split into multiple RTP packets.
+   */
+  private boolean gotFirstPacketOfH263Frame;
 
   /** Creates an instance. */
   public RtpH263Reader(RtpPayloadFormat payloadFormat) {
@@ -76,7 +84,10 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
   }
 
   @Override
-  public void onReceivingFirstPacket(long timestamp, int sequenceNumber) {}
+  public void onReceivingFirstPacket(long timestamp, int sequenceNumber) {
+    checkState(firstReceivedTimestamp == C.TIME_UNSET);
+    firstReceivedTimestamp = timestamp;
+  }
 
   @Override
   public void consume(
@@ -103,6 +114,12 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
     }
 
     if (pBitIsSet) {
+      if (gotFirstPacketOfH263Frame && fragmentedSampleSizeBytes > 0) {
+        // Received new H263 fragment, output data of previous fragment to decoder.
+        outputSampleMetadataForFragmentedPackets();
+      }
+      gotFirstPacketOfH263Frame = true;
+
       int payloadStartCode = data.peekUnsignedByte() & 0xFC;
       // Packets that begin with a Picture Start Code(100000). Refer RFC4629 Section 6.1.
       if (payloadStartCode < PICTURE_START_CODE) {
@@ -113,10 +130,10 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
       data.getData()[currentPosition] = 0;
       data.getData()[currentPosition + 1] = 0;
       data.setPosition(currentPosition);
-    } else {
+    } else if (gotFirstPacketOfH263Frame) {
       // Check that this packet is in the sequence of the previous packet.
       int expectedSequenceNumber = RtpPacket.getNextSequenceNumber(previousSequenceNumber);
-      if (sequenceNumber != expectedSequenceNumber) {
+      if (sequenceNumber < expectedSequenceNumber) {
         Log.w(
             TAG,
             Util.formatInvariant(
@@ -125,6 +142,12 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
                 expectedSequenceNumber, sequenceNumber));
         return;
       }
+    } else {
+      Log.w(
+          TAG,
+          "First payload octet of the H263 packet is not the beginning of a new H263 partition,"
+              + " Dropping current packet.");
+      return;
     }
 
     if (fragmentedSampleSizeBytes == 0) {
@@ -141,20 +164,10 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
     // Write the video sample.
     trackOutput.sampleData(data, fragmentSize);
     fragmentedSampleSizeBytes += fragmentSize;
+    fragmentedSampleTimeUs = toSampleUs(startTimeOffsetUs, timestamp, firstReceivedTimestamp);
 
     if (rtpMarker) {
-      if (firstReceivedTimestamp == C.TIME_UNSET) {
-        firstReceivedTimestamp = timestamp;
-      }
-      long timeUs = toSampleUs(startTimeOffsetUs, timestamp, firstReceivedTimestamp);
-      trackOutput.sampleMetadata(
-          timeUs,
-          isKeyFrame ? C.BUFFER_FLAG_KEY_FRAME : 0,
-          fragmentedSampleSizeBytes,
-          /* offset= */ 0,
-          /* cryptoData= */ null);
-      fragmentedSampleSizeBytes = 0;
-      isKeyFrame = false;
+      outputSampleMetadataForFragmentedPackets();
     }
     previousSequenceNumber = sequenceNumber;
   }
@@ -167,8 +180,8 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
   }
 
   /**
-   * Parses and set VOP Coding type and resolution. The {@link ParsableByteArray#position} is
-   * preserved.
+   * Parses and set VOP Coding type and resolution. The {@linkplain ParsableByteArray#getPosition()
+   * position} is preserved.
    */
   private void parseVopHeader(ParsableByteArray data, boolean gotResolution) {
     // Picture Segment Packets (RFC4629 Section 6.1).
@@ -209,6 +222,25 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
     }
     data.setPosition(currentPosition);
     isKeyFrame = false;
+  }
+
+  /**
+   * Outputs sample metadata of the received fragmented packets.
+   *
+   * <p>Call this method only after receiving an end of a H263 partition.
+   */
+  private void outputSampleMetadataForFragmentedPackets() {
+    checkNotNull(trackOutput)
+        .sampleMetadata(
+            fragmentedSampleTimeUs,
+            isKeyFrame ? C.BUFFER_FLAG_KEY_FRAME : 0,
+            fragmentedSampleSizeBytes,
+            /* offset= */ 0,
+            /* cryptoData= */ null);
+    fragmentedSampleSizeBytes = 0;
+    fragmentedSampleTimeUs = C.TIME_UNSET;
+    isKeyFrame = false;
+    gotFirstPacketOfH263Frame = false;
   }
 
   private static long toSampleUs(